active card
The description is unclear in the question how should the grid behave
when ‘fresh’ numbers and/or the active number and/or highlighted
numbers are clicked. the question mentions clicked for a second time so
I’ve gone with clicking the ‘active card’ (the card that was clicked to
originally) resets the board to a blank state but clicking any other
number resets the board with multiples of the new active card.